Add navigation menu and snake page
