+++ dist-tests/pages/SnakePage.js
@@ -1,0 +1,5 @@
+import { jsx as _jsx, jsxs as _jsxs } from "react/jsx-runtime";
+const SnakePage = () => {
+    return (_jsxs("div", { className: "snakePage", children: [_jsx("h2", { children: "Snake" }), _jsx("p", { children: "Coming soon..." })] }));
+};
+export default SnakePage;
