+++ src/App.css
@@ -48,6 +48,23 @@ header h1 {
   transform: translateY(-50%);
 }
 
+nav.mainNav {
+  position: absolute;
+  right: 1rem;
+  top: 50%;
+  transform: translateY(-50%);
+}
+
+nav.mainNav a {
+  margin-left: 1rem;
+  color: inherit;
+  text-decoration: none;
+}
+
+nav.mainNav a:hover {
+  text-decoration: underline;
+}
+
 @media (max-width: 768px) {
   header h1 {
     font-size: 1.5rem;
