+++ src/pages/SnakePage.tsx
@@ -1,0 +1,10 @@
+const SnakePage = () => {
+  return (
+    <div className="snakePage">
+      <h2>Snake</h2>
+      <p>Coming soon...</p>
+    </div>
+  )
+}
+
+export default SnakePage
